refactor: reorganize imports and remove backward compatibility code
