+++ src/diwire/__init__.py
@@ -1,5 +1,6 @@
-from diwire.container import Container, ScopedContainer
+from diwire.container import Container
 from diwire.container_context import container_context
+from diwire.container_scopes import ScopedContainer
 from diwire.service_key import Component
 from diwire.types import Injected, Lifetime
